Move gson out of the model
