+++ src/main/java/dev/ai4j/openai4j/Json.java
@@ -2,6 +2,7 @@ package dev.ai4j.openai4j;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+import dev.ai4j.openai4j.chat.Role;
 
 import static com.google.gson.FieldNamingPolicy.LOWER_CASE_WITH_UNDERSCORES;
 import static dev.ai4j.openai4j.MessageTypeAdapter.MESSAGE_TYPE_ADAPTER_FACTORY;
@@ -11,6 +12,7 @@ class Json {
     static final Gson GSON = new GsonBuilder()
             .setFieldNamingPolicy(LOWER_CASE_WITH_UNDERSCORES)
             .registerTypeAdapterFactory(MESSAGE_TYPE_ADAPTER_FACTORY)
+            .registerTypeAdapter(Role.class, new RoleAdapter())
             .create();
 
     static String toJson(Object o) {
