+++ src/test/java/dev/ai4j/openai4j/chat/ChatCompletionTest.java
@@ -1,5 +1,6 @@
 package dev.ai4j.openai4j.chat;
 
+import dev.ai4j.openai4j.FunctionCallUtil;
 import dev.ai4j.openai4j.OpenAiClient;
 import dev.ai4j.openai4j.RateLimitAwareTest;
 import org.junit.jupiter.api.Test;
@@ -95,12 +96,12 @@ class ChatCompletionTest extends RateLimitAwareTest {
         assertThat(functionCall.name()).isEqualTo("get_current_weather");
         assertThat(functionCall.arguments()).isNotBlank();
 
-        Map<String, Object> arguments = functionCall.argumentsAsMap();
+        Map<String, Object> arguments = FunctionCallUtil.argumentsAsMap(functionCall.arguments());
         assertThat(arguments).hasSize(1);
         assertThat(arguments.get("location").toString()).contains("Boston");
 
-        String location = functionCall.argument("location");
-        String unit = functionCall.argument("unit");
+        String location = FunctionCallUtil.argument(functionCall, "location");
+        String unit = FunctionCallUtil.argument(functionCall, "unit");
 
         String weatherApiResponse = getCurrentWeather(location, unit == null ? null : Unit.valueOf(unit));
